refactor: restructure and rename files
