+++ .gitignore
@@ -46,3 +46,4 @@ app.*.map.json
 /android/app/release
 
 .vscode/
+coverage/
