+++ lib/model/search_query_model.dart
