+++ lib/widgets/lyrics_page_widget.dart
@@ -1,4 +1,4 @@
-import 'package:ateez_lyrics/model/song.dart';
+import 'package:ateez_lyrics/model/song_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_widget_from_html_core/flutter_widget_from_html_core.dart';
